refactor: TDET-50 Separated Python operators from DAG. Refactored constants to be in a separate file
